requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0

# ML & Embeddings
sentence-transformers==3.4.1
faiss-cpu==1.8.0
torch==2.6.0
transformers==4.51.3
numpy==1.26.4
scikit-learn==1.5.2

# Data Processing
pandas==2.2.3
datasets==3.0.1
tqdm==4.66.5
langdetect==1.0.9

# LLM Integration
groq==0.11.0
httpx==0.27.2
tenacity==9.0.0

# Caching
cachetools==5.5.0
redis==5.1.1  # optional, falls back to in-memory

# Evaluation
nltk==3.9.1
rouge-score==0.1.2
sacrebleu==2.4.3

# Monitoring & Logging
structlog==24.4.0
prometheus-client==0.21.0

# API Utilities
python-multipart==0.0.12
python-jose[cryptography]==3.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Visualization
matplotlib==3.9.2
seaborn==0.13.2


typing_extensions>=4.10.0